streamlit/brasileirao/requirements.txt
numpy==1.23.4
pandas==1.5.1
plotly==5.11.0
streamlit==1.24.0
Pillow==9.5.0
beautifulsoup4==4.12.2